Lesson 3: Advanced solution of hard task in lesson 3, now with large numbers

+++ src/lesson3/task1/Loop.kt
@@ -201,14 +201,14 @@ fun hasDifferentDigits(n: Int): Boolean = TODO()
 fun squareSequenceDigit(n: Int): Int {
     var position = 0
     var natural = 1
-    var square: Int
+    var square: Long
     do {
-        square = natural * natural
+        square = (natural * natural).toLong()
         val digitCount = (log10(square.toDouble()) + 1).toInt()
         position += digitCount
         natural++
     } while (position < n)
-    return (square / (10.0.pow(position - n)).toInt()) % 10
+    return (square / (10.0.pow(position - n))).toInt() % 10
 }
 
 /**
